refactoring first iteration
