+++ .github/workflows/test.yml
@@ -79,4 +79,46 @@ jobs:
       - run: pip install -r requirements.txt
       - run: pip3 install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cpu
       - run: pip install pyg-lib torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric -f https://data.pyg.org/whl/torch-1.13.0+cpu.html
-      - run: ./scripts/local_test_ssl_pipeline.sh
+      - run: ./scripts/local_test_pyg_ssl.sh
+
+  pyg-delphes-pipeline:
+    runs-on: ubuntu-20.04
+    needs: [deps-pyg]
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: '3.8.10'
+          cache: 'pip'
+      - run: pip install -r requirements.txt
+      - run: pip3 install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cpu
+      - run: pip install pyg-lib torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric -f https://data.pyg.org/whl/torch-1.13.0+cpu.html
+      - run: ./scripts/local_test_pyg_delphes.sh
+
+  pyg-cms-pipeline:
+    runs-on: ubuntu-20.04
+    needs: [deps-pyg]
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: '3.8.10'
+          cache: 'pip'
+      - run: pip install -r requirements.txt
+      - run: pip3 install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cpu
+      - run: pip install pyg-lib torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric -f https://data.pyg.org/whl/torch-1.13.0+cpu.html
+      - run: ./scripts/local_test_pyg_cms.sh
+
+  pyg-clic-pipeline:
+    runs-on: ubuntu-20.04
+    needs: [deps-pyg]
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: '3.8.10'
+          cache: 'pip'
+      - run: pip install -r requirements.txt
+      - run: pip3 install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cpu
+      - run: pip install pyg-lib torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric -f https://data.pyg.org/whl/torch-1.13.0+cpu.html
+      - run: ./scripts/local_test_pyg_clic.sh
