+++ mlpf/pyg/get_data_clic.sh
