+++ mlpf/pyg/get_data_delphes.sh
