--- mlpf/pyg_ssl/environment.yml
@@ -1,20 +1,0 @@
-name: mlpf
-channels:
-    - pyg
-    - pytorch
-    - conda-forge
-dependencies:
-    - python=3.9
-    - numpy
-    - pandas
-    - scikit-learn
-    - matplotlib
-    - pytorch
-    - pyg
-    - cpuonly
-    - mplhep
-    - tqdm
-    - autopep8
-    - pip
-    - pip:
-      - jupyter-book
